🧬 Rename the module to be more clear meaning.
🧬 Refactor the code usage with the new module naming.
