+++ pymock_api/runner.py
@@ -5,13 +5,13 @@ from pathlib import Path
 from typing import List, Optional
 
 try:
-    from pymock_api.command.cmd_ps import CommandProcessor, dispatch_command_processor
+    from pymock_api.command.process import CommandProcessor, dispatch_command_processor
     from pymock_api.model import ParserArguments
 except (ImportError, ModuleNotFoundError):
     runner_dir = os.path.dirname(os.path.abspath(__file__))
     path = str(Path(runner_dir).parent.absolute())
     sys.path.append(path)
-    from pymock_api.command.cmd_ps import CommandProcessor, dispatch_command_processor
+    from pymock_api.command.process import CommandProcessor, dispatch_command_processor
     from pymock_api.model import ParserArguments
