+++ test/unit_test/command/process.py
@@ -11,7 +11,8 @@ from unittest.mock import MagicMock, Mock, call, patch
 import pytest
 
 from pymock_api._utils.file_opt import YAML
-from pymock_api.command.cmd_ps import (
+from pymock_api.command.options import SubCommand, get_all_subcommands
+from pymock_api.command.process import (
     BaseCommandProcessor,
     NoSubCmd,
     SubCmdCheck,
@@ -21,7 +22,6 @@ from pymock_api.command.cmd_ps import (
     make_command_chain,
     run_command_chain,
 )
-from pymock_api.command.options import SubCommand, get_all_subcommands
 from pymock_api.model import (
     ParserArguments,
     SubcmdCheckArguments,
@@ -250,7 +250,7 @@ class TestNoSubCmd(BaseCommandProcessorTestSpec):
         command = _given_command(app_type="Python web library")
         command.run = MagicMock()
 
-        with patch("pymock_api.command.cmd_ps.YAML", return_value=FakeYAML) as mock_instantiate_writer:
+        with patch("pymock_api.command.process.YAML", return_value=FakeYAML) as mock_instantiate_writer:
             with patch.object(WSGIServer, "generate", return_value=command) as mock_sgi_generate:
                 cmd_ps(mock_parser_arg)
                 mock_sgi_generate.assert_not_called()
@@ -313,7 +313,7 @@ class TestSubCmdRun(BaseCommandProcessorTestSpec):
         command = _given_command(app_type="Python web library")
         command.run = MagicMock()
 
-        with patch("pymock_api.command.cmd_ps.YAML", return_value=FakeYAML) as mock_instantiate_writer:
+        with patch("pymock_api.command.process.YAML", return_value=FakeYAML) as mock_instantiate_writer:
             with patch.object(ASGIServer, "generate", return_value=command) as mock_asgi_generate:
                 with patch.object(WSGIServer, "generate", return_value=command) as mock_wsgi_generate:
                     if should_raise_exc:
@@ -356,7 +356,7 @@ class TestSubCmdRun(BaseCommandProcessorTestSpec):
     def _expected_argument_type(self) -> Type[SubcmdRunArguments]:
         return SubcmdRunArguments
 
-    @patch("pymock_api.command.cmd_ps.import_web_lib.auto_ready", return_value=None)
+    @patch("pymock_api.command.process.import_web_lib.auto_ready", return_value=None)
     def test_auto_with_nonexist_lib(self, mock_auto_ready: Mock, cmd_ps: SubCmdRun):
         with pytest.raises(RuntimeError) as exc_info:
             cmd_ps._initial_server_gateway(lib=_Test_Auto_Type)
@@ -441,7 +441,7 @@ class TestSubCmdConfig(BaseCommandProcessorTestSpec):
         )
 
         with patch("builtins.print", autospec=True, side_effect=print) as mock_print:
-            with patch("pymock_api.command.cmd_ps.YAML", return_value=FakeYAML) as mock_instantiate_writer:
+            with patch("pymock_api.command.process.YAML", return_value=FakeYAML) as mock_instantiate_writer:
                 cmd_ps(mock_parser_arg)
 
                 mock_instantiate_writer.assert_called_once()
@@ -484,7 +484,7 @@ class TestSubCmdConfig(BaseCommandProcessorTestSpec):
         )
 
         # Run target function to test
-        with patch("pymock_api.command.cmd_ps.YAML", return_value=FakeYAML) as mock_instantiate_writer:
+        with patch("pymock_api.command.process.YAML", return_value=FakeYAML) as mock_instantiate_writer:
             with pytest.raises(AssertionError) as exc_info:
                 cmd_ps.process(invalid_args)
 
@@ -669,7 +669,7 @@ class TestSubCmdCheck(BaseCommandProcessorTestSpec):
         mock_parser_arg = _given_parser_args(
             subcommand=_Test_SubCommand_Check, swagger_doc_url=_Swagger_API_Document_URL, stop_if_fail=stop_if_fail
         )
-        with patch("pymock_api.command.cmd_ps.load_config") as mock_load_config:
+        with patch("pymock_api.command.process.load_config") as mock_load_config:
             mock_load_config.return_value = load_config(dummy_yaml_path)
             with patch.object(SubCmdCheck, "_get_swagger_config") as mock_get_swagger_config:
                 with open(api_resp_path, "r", encoding="utf-8") as file_stream:
@@ -714,7 +714,7 @@ class TestSubCmdCheck(BaseCommandProcessorTestSpec):
 
     @pytest.mark.parametrize("swagger_config_response", RESPONSE_JSON_PATHS)
     def test__get_swagger_config(self, swagger_config_response: str, cmd_ps: SubCmdCheck):
-        with patch("pymock_api.command.cmd_ps.URLLibHTTPClient.request") as mock_api_client_request:
+        with patch("pymock_api.command.process.URLLibHTTPClient.request") as mock_api_client_request:
             with open(swagger_config_response, "r", encoding="utf-8") as file_stream:
                 mock_api_client_request.return_value = json.loads(file_stream.read())
 
@@ -781,6 +781,6 @@ def test_make_command_chain_if_duplicated_subcmd():
     assert re.search(r"subcommand.{1,64}has been used", str(exc_info.value), re.IGNORECASE)
 
     # Remove the invalid object for test could run finely.
-    from pymock_api.command.cmd_ps import _COMMAND_CHAIN
+    from pymock_api.command.process import _COMMAND_CHAIN
 
     _COMMAND_CHAIN.pop(-1)
